chore(playbook): update router_dhcp.yml
Update the router_dhcp.yml file to save the MAC addresses of the DNS, webserver, and db to a file, install DHCP, copy new DHCP configuration to remote host, and restart the DHCP service to apply the changes.

MAC addresses are now saved to a file for future reference and the installation of DHCP has been updated to use dnf instead of yum.

The DHCP configuration has been updated using a template and the DHCP service is now enabled to start on boot.